requirements.txt
google-cloud-bigquery==3.34.0
google-cloud-storage==2.19.0
pandas==2.2.3
python-dotenv==1.1.1
spotipy==2.25.1